Move Report Header into Component
Moving this into a standalone component as a first step to moving it
inside of each report.
