+++ packages/frontend/app/components/reports/subject.hbs
@@ -9,53 +9,14 @@
         {{t "general.backToReports"}}
       </LinkTo>
     </div>
-    <div class="header" data-test-report-header>
-      <h2 class="title" data-test-report-title>
-        {{#if this.reportTitleData.isResolved}}
-          <EditableField
-            data-test-title
-            @value={{this.reportTitle}}
-            @save={{perform this.changeTitle}}
-            @close={{this.revertTitleChanges}}
-            @saveOnEnter={{true}}
-            @showIcon={{true}}
-            @closeOnEscape={{true}} as |isSaving|
-          >
-            <input
-              aria-label={{t "general.reportTitle"}}
-              type="text"
-              value={{this.reportTitle}}
-              placeholder={{this.constructedReportTitle}}
-              disabled={{isSaving}}
-              {{on "input" (pick "target.value" (set this "title"))}}
-              {{on "keyup" (fn this.addErrorDisplayFor "title")}}
-            >
-            <ValidationError @errors={{get-errors-for this.title}} />
-          </EditableField>
-        {{/if}}
-      </h2>
-      <div class="actions">
-        <button
-          type="button"
-          disabled={{or (not this.dataIsLoaded) this.downloadReport.isRunning}}
-          {{on "click" (perform this.downloadReport)}}
-          data-test-download
-        >
-          {{#if this.finishedBuildingReport}}
-            <FaIcon @icon="check" />
-          {{else if (or (not this.dataIsLoaded) this.downloadReport.isRunning)}}
-            <LoadingSpinner />
-          {{else}}
-            <FaIcon @icon="download" />
-          {{/if}}
-          {{t "general.downloadResults"}}
-        </button>
-      </div>
-    </div>
+    <Reports::Subject::Header
+      @report={{@report}}
+      @school={{this.school}}
+      @year={{@year}}
+    />
     {{#let (load this.reportDescriptionPromise) as |p|}}
       {{#if p.isResolved}}
         <Reports::SubjectResults
-          @setDataIsLoaded={{this.setDataIsLoaded}}
           @subject={{@report.subject}}
           @prepositionalObject={{@report.prepositionalObject}}
           @prepositionalObjectTableRowId={{@report.prepositionalObjectTableRowId}}
