+++ packages/frontend/app/styles/components.scss
@@ -140,6 +140,7 @@
 
 @import "components/reports/new-subject";
 @import "components/reports/subject";
+@import "components/reports/subject-header";
 @import "components/reports/subjects";
 @import "components/reports/root";
 @import "components/reports/list-loading";
