+++ packages/frontend/app/styles/components/reports/subject-header.scss
@@ -1,0 +1,27 @@
+@use "../../ilios-common/mixins" as cm;
+@use "../../ilios-common/colors" as c;
+@use "../../ilios-common/constants";
+
+.reports-subject-header {
+  border-bottom: 1px solid c.$culturedGrey;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  padding: calc(constants.$golden-ratio-large * 1rem) 0
+    calc(constants.$golden-ratio-small * 1rem) 0;
+  margin: 0 0.8rem;
+
+  @include cm.for-tablet-and-up {
+    flex-direction: row;
+  }
+
+  h2 {
+    @include cm.ilios-heading-h4;
+    flex-grow: 2;
+
+    .is-editing,
+    input {
+      width: 90%;
+    }
+  }
+}
